Implement doubleclick handling in details-table
The doubleclick switches focus to our details table.
This focus change leads to a loss of our
current selection in the previously focused combobox.
Inserting the doubleclicked item now would not remove the previously selected text from the combobox.
We fix that by saving and restoring the combobox selection state on focus change events.